% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.PPC.binom.exact.R
\name{check.PPC.binom.exact}
\alias{check.PPC.binom.exact}
\title{Argument Check}
\usage{
check.PPC.binom.exact(n, alpha1, beta1, alpha2, beta2, alphaD, betaD, xi, v, q, M, plot)
}
\value{
An error message.
}
\description{
Checks if the arguments of the \code{\link{PPC.binom.exact}} function are correctly specified.
}
\note{
This is an internal function. Cannot be used directly.
}
\author{
Michele Cianfriglia \email{michele.cianfriglia@uniroma1.it}
}
\keyword{internal}
