% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.boolean.R
\name{is.boolean}
\alias{is.boolean}
\title{Boolean}
\usage{
is.boolean(x)
}
\value{
A logical value.
}
\description{
Checks if an object is a boolean.
}
\note{
This is an internal function. Cannot be used directly.
}
\author{
Michele Cianfriglia \email{michele.cianfriglia@uniroma1.it}
}
\keyword{internal}
