% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.number.R
\name{is.number}
\alias{is.number}
\title{Number}
\usage{
is.number(x, int = FALSE)
}
\value{
A logical value.
}
\description{
Checks if an object is a number. If required, checks that the object is an integer.
}
\note{
This is an internal function. Cannot be used directly.
}
\author{
Michele Cianfriglia \email{michele.cianfriglia@uniroma1.it}
}
\keyword{internal}
