% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEC.binom.exact.R
\name{PEC.binom.exact}
\alias{PEC.binom.exact}
\title{Predictive Expectation Criterion: Binomial Model

\loadmathjax{}}
\usage{
PEC.binom.exact(n, alpha1, beta1, alpha2, beta2, alphaD, betaD, v,
                q = 1, M = 2500, plot = FALSE, seed = NULL)
}
\arguments{
\item{n}{The sample size. A vector of positive integers arranged in ascending order.}

\item{alpha1, beta1}{Parameters of the first Beta distribution.}

\item{alpha2, beta2}{Parameters of the second Beta distribution.}

\item{alphaD, betaD}{Parameters of the design Beta distribution.}

\item{v}{A value in \mjseqn{(0, 1)}.}

\item{q}{The order of the Wasserstein distance. Must be a value in \mjseqn{[1, \infty)}.}

\item{M}{The number of Monte Carlo replications. Must be a positive integer.}

\item{plot}{Logical. If \code{TRUE}, a plot shows the behavior of the predictive expectation as a function of the sample size.}

\item{seed}{A number or \code{NULL}. If \code{NULL} (default), the user does not specify any seed.}
}
\value{
A list with the following components \cr
\item{pred.exp}{The predictive expectation.}
\item{t.opt}{The optimal threshold.}
\item{n.opt}{The optimal sample size.}
}
\description{
Implements the predictive expectation criterion in the conjugate Binomial model using the exact expression of the \mjseqn{\ell_1} Wasserstein distance of order \mjseqn{q}.
}
\details{

}
\note{
If the arguments of the \code{PEC.binom.exact} function are not correctly specified, the execution is stopped and an error message is generated.
}
\examples{

}
\references{
M. Cianfriglia, T. Padellini and P. Brutti (2022) \emph{Wasserstein consensus for Bayesian sample size determination}
}
\seealso{
\code{\link{PPC.binom.exact}}
}
\author{
Michele Cianfriglia \email{michele.cianfriglia@uniroma1.it}
}
