% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.exp.binom.exact.R
\name{pred.exp.binom.exact}
\alias{pred.exp.binom.exact}
\title{Predictive Expectation}
\usage{
pred.exp.binom.exact(n, alpha1, beta1, alpha2, beta2, alphaD, betaD, q, M)
}
\value{
A numeric value.
}
\description{
Computes the predictive expectation for the \code{\link{PEC.binom.exact}} function.
}
\note{
This is an internal function. Cannot be used directly.
}
\author{
Michele Cianfriglia \email{michele.cianfriglia@uniroma1.it}
}
\keyword{internal}
