% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.prob.binom.exact.R
\name{pred.prob.binom.exact}
\alias{pred.prob.binom.exact}
\title{Predictive Probability}
\usage{
pred.prob.binom.exact(n, alpha1, beta1, alpha2, beta2, alphaD, betaD, xi, q, M)
}
\value{
A numeric value.
}
\description{
Computes the predictive probability for the \code{\link{PPC.binom.exact}} function.
}
\note{
This is an internal function. Cannot be used directly.
}
\author{
Michele Cianfriglia \email{michele.cianfriglia@uniroma1.it}
}
\keyword{internal}
